+++ src/Utils/CmsUtils.php
@@ -6,16 +6,16 @@ namespace Basecom\FixturePlugin\Utils;
 
 use Shopware\Core\Content\Cms\CmsPageEntity;
 use Shopware\Core\Framework\Context;
-use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsAnyFilter;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 
 class CmsUtils
 {
-    private EntityRepositoryInterface $cmsPageRepository;
+    private EntityRepository $cmsPageRepository;
 
-    public function __construct(EntityRepositoryInterface $cmsPageRepository)
+    public function __construct(EntityRepository $cmsPageRepository)
     {
         $this->cmsPageRepository = $cmsPageRepository;
     }