+++ src/Utils/PaymentMethodUtils.php
@@ -7,15 +7,27 @@ namespace Basecom\FixturePlugin\Utils;
 use Shopware\Core\Checkout\Payment\Cart\PaymentHandler\InvoicePayment;
 use Shopware\Core\Checkout\Payment\PaymentMethodEntity;
 use Shopware\Core\Framework\Context;
-use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 
 class PaymentMethodUtils
 {
-    private EntityRepositoryInterface $paymentMethodRepository;
+    /**
+     * @TODO Replace with real typehint when Shopware 6.4 support is dropped.
+     *
+     * @var EntityRepository
+     */
+    private $paymentMethodRepository;
 
-    public function __construct(EntityRepositoryInterface $paymentMethodRepository)
+    /**
+     * @TODO Replace with real typehint when Shopware 6.4 support is dropped.
+     *
+     * @param EntityRepository $paymentMethodRepository
+     *
+     * @noinspection PhpMissingParamTypeInspection We can not use typehints until PaymentMethodRepositoryDecorator is removed (Shopware 6.5).
+     */
+    public function __construct($paymentMethodRepository)
     {
         $this->paymentMethodRepository = $paymentMethodRepository;
     }