+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.1] - 2023-05-26
+### Fixed
+- Restored compatibility for Shopware 6.4 by removing typehints for **EntityRepository** in util classes.
+
+## [2.2.0] - 2023-05-09
+### Added
+- Added tests for PHP 8.2
+
+### Changed
+- Switched from **EntityRepositoryInterface** to **EntityRepository** in util classes to support Shopware 6.5
+
+## Removed
+- Dropped support for PHP 7.4 and 8.0
+
 ## [2.1.0] - 2022-09-19
 ### Added
 - Added the option `--with-dependencies` / `-w` to the `fixture:run:single` command
@@ -44,7 +58,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.6.0] - 2022-05-09
 ### Added
 - Added more helper functions:
-  - `getMediaDefaultFolderId` 
+  - `getMediaDefaultFolderId`
 
 ## [1.5.0] - 2022-04-20
 ### Added
@@ -110,6 +124,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `getNotSpecifiedSalutationId`
   - `getGermanCountryId`
 
+[2.2.1]: https://github.com/basecom/FixturesPlugin/compare/2.2.0...2.2.1
+[2.2.0]: https://github.com/basecom/FixturesPlugin/compare/2.1.0...2.2.0
 [2.1.0]: https://github.com/basecom/FixturesPlugin/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/basecom/FixturesPlugin/compare/1.8.0...2.0.0
 [1.8.0]: https://github.com/basecom/FixturesPlugin/compare/1.7.0...1.8.0